Algorithm update
Since throttle has a value between 1000 to 2000 for both directions, the
previous method is redundant.

A new definition "bottom" is the threshold for the throttle possition.
This method will switch between direction with recpect to channel_5 value.

The steering is now using an OR gate to map the appropriate value of z.