Time conversion
The values read from each channel is in ms and it needed
to ne converted to a base 2 number for easier arithmetic operation.

The values are ranged from 900ms to 2000ms; With some trim settings on the
transmitter values from 1000ms to 2000ms worked best. These will be mapped
to values from 0 to 255. (2^8 (neg 1 to conpencate for zero)).

Fine trim needs to be made according to the car it's used for.