Change PCINT registers to acomodate for Nano
To acheive the same result as Arduino Mega, The intrupt enabled pins
nedded to be changed. Acording to the chip datasheet, Analog pins5:0 are
corresponding to PCINT13:8 which are under PCI1 pin change intrrupt group
and are masked by PCMSK1 register.

To enable intrrupts on PCINT13:8; [PCINT15:8] and PCIE1 bit in PCICR
should be set.

Pin Change Mask Register 1 [PCMSK1] has an offset of 0x6c and Reset of
0x00